+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* re-write `WP_Beta_Tester::get_current_wp_release()` to check https://api.wordpress.org/core/stable-check/1.0/
 
 #### 3.0.9 / 2020-12-01
 * add conditional for filter to fix `core_update_footer()`, fixed in [r49708](https://core.trac.wordpress.org/changeset/49708)